+++ test/support/index.js
@@ -5,3 +5,10 @@ import * as assertUtils from './assert'
 export {sinon}
 
 export const assert = Object.assign({}, assertNode, sinon.assert, assertUtils)
+
+
+export function collectEvents (obs) {
+  const events = []
+  obs.onValue((v) => events.unshift(v))
+  return events
+}